+++ print/src/main.rs
@@ -94,9 +94,7 @@ fn main() -> anyhow::Result<()> {
         return print_version();
     }
 
-    let log_format = log_format.as_ref().map(|s| s.as_str());
-
-    defmt_decoder::log::init_logger(log_format, json, move |metadata| match verbose {
+    defmt_decoder::log::init_logger(log_format.as_deref(), json, move |metadata| match verbose {
         false => defmt_decoder::log::is_defmt_frame(metadata), // We display *all* defmt frames, but nothing else.
         true => true,                                          // We display *all* frames.
     });