+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- [#765]: `defmt-decoder`: Add support for customizable logger formatting
+
+[#765]: https://github.com/knurling-rs/defmt/pull/765
 
 - [#766] `decoder::log`: Rename `PrettyLogger` to `StdoutLogger`